+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono, Raleway } from "next/font/google";
 import "./globals.css";
-import Top from "./components/top";
+import Top from "../components/ui/top";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
