+++ types/pages.ts
